requirements.txt
pyyaml>=6.0.1
pytest>=7.4.0
python-dotenv>=1.0.0
Jinja2>=3.1.2 